ISSUE #2169 - Move activities filtering logic to selector
